action update order

+++ src/helpers/axios.js
@@ -23,9 +23,9 @@ axiosInstance.interceptors.request.use((req) => {
 axiosInstance.interceptors.response.use((res) => {
     return res;
 }, (error) => {
-    console.log(error);
-    const { status } = error.response;
-    if(status === 500){
+    console.log(error.response);
+    const status = error.response ? error.response.status : 500;
+    if(status && status === 500){
         localStorage.clear();
         store.dispatch({ type: authConstants.LOGOUT_SUCCESS });
     }
